+++ src/app/decoration/page.tsx
@@ -1,7 +1,11 @@
 import DecorationForms from "@/components/decoration/DecorationForms";
 
 const DecorationPage = () => {
-    return <DecorationForms />;
+    return (
+        <div className="inner">
+            <DecorationForms />
+        </div>
+    );
 };
 
 export default DecorationPage;
