+++ src/components/common/index.ts
@@ -1,2 +1,3 @@
 export { default as InputForm } from "./InputForm";
 export { default as Button } from "./Button";
+export { default as Carousel } from "./Carousel";
